Add example data for ingest and phylogenetic CI workflows [#2]
This is the full NCBI yellow-fever dataset (taxon ID 11089) as
downloaded on 21 Nov 2024, and the ingested sequences and metadata as
downloaded from Nextstrain on 21 Nov 2024.

CI _could_ work just fine without embedding this data in the repo, but
having a hard-coded example-data file means that we don't depend on
NCBI or Nextstrain.org for successful CI runs.